Fix the earliest date to avoid breaking unit tests
It's convenient to have dates in unit tests. The moving earliest date
eventually broke the unit tests here and could conceivably break them
elsewhere. Rather than have the tests break when building on a random
day, make them break when the hardcoded default earliest date changes.
This should only affect users on package updates.

A side benefit of this is that it will reduce a source of nondeterminism
when building. It's arguable that the end date should also be fixed for
the same reason, but that's not addressed here.